+++ common/gcmage.lua
@@ -216,10 +216,10 @@ local restingMaxMP = false
 
 local lag = false
 
-local WeaponOverrideTable = {
-    [1] = '1',
-    [2] = '2',
-    [3] = '3',
+local WeaponOverrideIndexes = {
+    ['1'] = 1,
+    ['2'] = 2,
+    ['3'] = 3,
 }
 
 local weapon_override = 1
@@ -323,11 +323,18 @@ function gcmage.DoCommands(args, sets)
         lag =  not lag
         gcinclude.Message('[Note: Midcast Delays are disabled if Lag is true] Lag', lag)
     elseif ((args[1] == 'weapon' or args[1] == 'wl') and player.MainJob ~= 'BLM') then
-        weapon_override = weapon_override + 1
-        if (weapon_override > #WeaponOverrideTable) then
-            weapon_override = 1
+        if (args[2] ~= nil) then
+            local cycleIndex = WeaponOverrideIndexes[args[2]]
+            if (cycleIndex ~= nil) then
+                weapon_override = cycleIndex
+            end
+        else
+            weapon_override = weapon_override + 1
+            if (weapon_override > #WeaponOverrideIndexes) then
+                weapon_override = 1
+            end
         end
-        gcinclude.Message('Weapon Loadout', WeaponOverrideTable[weapon_override])
+        gcinclude.Message('Weapon Loadout', tostring(weapon_override))
     end
 
     if (player.MainJob == 'RDM' or player.MainJob == 'WHM') then
@@ -574,7 +581,7 @@ function gcmage.DoPrecast(sets, fastCastValue, cureCastMeritValue)
 
             if (whmYellow) then
                 if (gcdisplay.GetCycle('TP') ~= 'Off' and (player.Status == 'Engaged' or player.TP > 0)) then
-                    local weapon = sets['Weapon_Loadout_' .. WeaponOverrideTable[weapon_override]]
+                    local weapon = sets['Weapon_Loadout_' .. tostring(weapon_override)]
                     yellow = gFunc.Combine(yellow, weapon)
                 end
             end
@@ -637,7 +644,7 @@ function gcmage.SetupMidcastDelay(sets, fastCastValue, cureCastMeritValue)
                     local hpUp = sets.Cheat_HPUp
 
                     if (player.MainJob ~= 'BLM' and gcdisplay.GetCycle('TP') ~= 'Off' and (player.Status == 'Engaged' or player.TP > 0)) then
-                        local weapon = sets['Weapon_Loadout_' .. WeaponOverrideTable[weapon_override]]
+                        local weapon = sets['Weapon_Loadout_' .. tostring(weapon_override)]
                         hpDownC3 = gFunc.Combine(hpDownC3, weapon)
                         hpDownC4 = gFunc.Combine(hpDownC4, weapon)
                         hpUp = gFunc.Combine(hpUp, weapon)
@@ -688,7 +695,7 @@ function gcmage.SetupMidcastDelay(sets, fastCastValue, cureCastMeritValue)
 
             if (whmYellow) then
                 if (gcdisplay.GetCycle('TP') ~= 'Off' and (player.Status == 'Engaged' or player.TP > 0)) then
-                    local weapon = sets['Weapon_Loadout_' .. WeaponOverrideTable[weapon_override]]
+                    local weapon = sets['Weapon_Loadout_' .. tostring(weapon_override)]
                     yellow = gFunc.Combine(yellow, weapon)
                 end
             end
@@ -801,7 +808,7 @@ end
 function gcmage.EquipWeaponLoadout()
     local player = gData.GetPlayer()
     if (player.MainJob ~= 'BLM' and gcdisplay.GetCycle('TP') ~= 'Off' and (player.Status == 'Engaged' or player.TP > 0)) then
-        gFunc.EquipSet('Weapon_Loadout_' .. WeaponOverrideTable[weapon_override])
+        gFunc.EquipSet('Weapon_Loadout_' .. tostring(weapon_override))
     end
 end
 
@@ -932,7 +939,7 @@ function gcmage.SetupInterimEquipSet(sets, isRanged)
     if (gcdisplay.IdleSet == 'Override') then interimSet = sets.Override end
 
     if (player.MainJob ~= 'BLM' and gcdisplay.GetCycle('TP') ~= 'Off' and (player.Status == 'Engaged' or player.TP > 0)) then
-        interimSet = gFunc.Combine(interimSet, sets['Weapon_Loadout_' .. WeaponOverrideTable[weapon_override]])
+        interimSet = gFunc.Combine(interimSet, sets['Weapon_Loadout_' .. tostring(weapon_override)])
     end
 
     if (isRanged) then